refactor: frame 관련 패키지 생성
