+++ src/main/java/bowling/view/ResultView.java
@@ -1,6 +1,6 @@
 package bowling.view;
 
-import bowling.domain.Frames;
+import bowling.domain.frame.Frames;
 
 public interface ResultView {
     void print(Frames bowlingGame);
