+++ src/test/java/bowling/domain/FramesTest.java
@@ -1,5 +1,6 @@
 package bowling.domain;
 
+import bowling.domain.frame.Frames;
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
