+++ src/test/java/bowling/domain/NormalFrameTest.java
@@ -1,5 +1,7 @@
 package bowling.domain;
 
+import bowling.domain.frame.Frame;
+import bowling.domain.frame.NormalFrame;
 import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
